Refactor FilenamePassesIncludeExcludeFilter into tools package
