+++ tools/filetools.go
@@ -12,6 +12,8 @@ import (
 	"strings"
 )
 
+var localDirSet = NewStringSetFromSlice([]string{".", "./", ".\\"})
+
 // FileOrDirExists determines if a file/dir exists, returns IsDir() results too.
 func FileOrDirExists(path string) (exists bool, isDir bool) {
 	fi, err := os.Stat(path)
@@ -120,4 +122,66 @@ func VerifyFileHash(oid, path string) error {
 	}
 
 	return nil
+}
+
+// FilenamePassesIncludeExcludeFilter returns whether a given filename passes the include / exclude path filters
+// Only paths that are in includePaths and outside excludePaths are passed
+// If includePaths is empty that filter always passes and the same with excludePaths
+// Both path lists support wildcard matches
+func FilenamePassesIncludeExcludeFilter(filename string, includePaths, excludePaths []string) bool {
+	if len(includePaths) == 0 && len(excludePaths) == 0 {
+		return true
+	}
+
+	filename = filepath.Clean(filename)
+	if len(includePaths) > 0 {
+		matched := false
+		for _, inc := range includePaths {
+			inc = filepath.Clean(inc)
+
+			// Special case local dir, matches all (inc subpaths)
+			if _, local := localDirSet[inc]; local {
+				matched = true
+				break
+			}
+
+			matched, _ = filepath.Match(inc, filename)
+			if !matched {
+				// Also support matching a parent directory without a wildcard
+				if strings.HasPrefix(filename, inc+string(filepath.Separator)) {
+					matched = true
+				}
+			}
+
+			if matched {
+				break
+			}
+
+		}
+		if !matched {
+			return false
+		}
+	}
+
+	if len(excludePaths) > 0 {
+		for _, ex := range excludePaths {
+			ex = filepath.Clean(ex)
+
+			// Special case local dir, matches all (inc subpaths)
+			if _, local := localDirSet[ex]; local {
+				return false
+			}
+
+			if matched, _ := filepath.Match(ex, filename); matched {
+				return false
+			}
+
+			// Also support matching a parent directory without a wildcard
+			if strings.HasPrefix(filename, ex+string(filepath.Separator)) {
+				return false
+			}
+		}
+	}
+
+	return true
 }
